+++ CHANGELOG.md
@@ -5,6 +5,31 @@ All notable changes to PGS will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). Dates are *YYYY-MM-DD*.
 
+## **2.1** *(2025-xx-xx)*
+
+### Added
+* `smoothLaneRiesenfeld` to `PGS_Morphology`. Smooths a shape using Lane-Riesenfeld curve subdivision with 4-point refinement to reduce contraction.
+* Additional method signature for `PGS_Conversion.roundVertexCoords()` that accepts a number of decimal places.
+* `interiorAngles()` to `PGS_ShapePredicates`. Calculates all interior angles of a polygon.
+* `forEachShape()` and `forEachShapeWithIndex()`* to `PGS_Processing`. Applies a specified transformation function of a desired type `T` to each child of the given PShape, returning a list of  `T` (*additionally with child's index).
+* `maximumInscribedTriangle()` to `PGS_Optimisation`. Finds an approximate largest area triangle (of arbitrary orientation) contained within a polygon.
+* `closestPoint()` to `PGS_Optimisation`. Finds the closest point in a collection of points to a specified point.
+* `distanceTree()` to `PGS_Contour`. Generates a tree structure representing the shortest paths from a start point to all other vertices in a mesh.
+
+### Changes
+* Optimised `PGS_CirclePacking.tangencyPack()`. It's now around 1.5-2x faster and has higher precision.
+* `PGS_Conversion.roundVertexCoords()` now returns a rounded copy of the input (rather than mutating the input).
+* Outputs from `PGS_Conversion.toDualGraph()` will now always iterate deterministically on inputs with the same geometry but having a different structure.
+* `PGS_Contour.straightSkeleton()` now always uses a more robust approach (which has been sped up considerably too).
+
+### Fixed
+* `PGS_Morphology.rounding()` no longer gives invalid results.
+* `PGS_ShapePredicates.elongation()` now correctly measures shape elongation (previously inverted, now returns 1 for highly elongated shapes).
+* `PGS_Conversion.toGraph()` now processes `LINES` shapes correctly.
+* `PGS_Meshing.urquhartFaces()` no longer errors on triangulation inputs with no constraints.
+
+### Removed
+
 ## **2.0** *(2025-01-11)*
 
 **NOTE: Beginning at v2.0, PGS is built with Java 17.**